sounds and backgrounds
backgrounds of menu and end screen changed to a midday beach and a sunset beach shore. they somewhat hold meaning, sunset meaning the hunt for plastic is over, and midday meaning we're about to begin, with the ocean underwater background from before of course meaning that we're looking for plastic to clean the ocean. the ambient sound is kind of creepy, which I like since its horrifying what happens to the sea creatures from our plastic waste, so it has meaning. the line clear sound effect is just bubbles, just thought it fit in appropriately.